Library and makefile done
