+++ Assign5/client.c
@@ -1,4 +1,14 @@
-#include"library.c"
+#include <stdio.h>
+#include <unistd.h>
+#include <string.h>
+#include <sys/socket.h>
+#include <sys/types.h>
+#include <arpa/inet.h>
+#include <netinet/in.h>
+#include <errno.h>
+#include <pthread.h>
+#include <stdlib.h>
+
 int main()
 {
