+++ Assign5/server.c
@@ -1,5 +1,14 @@
+#include <stdio.h>
+#include <unistd.h>
+#include <string.h>
+#include <sys/socket.h>
+#include <sys/types.h>
+#include <arpa/inet.h>
+#include <netinet/in.h>
+#include <errno.h>
+#include <pthread.h>
+#include <stdlib.h>
 
-#include "library.c"
 int main()
 {
